Add user deletion functionality
This commit adds the ability to delete a user from the system. The UserController now includes a new DELETE endpoint that accepts a GUID as a parameter and deletes the corresponding user if found. If the user does not exist, a 404 Not Found response is returned. On successful deletion, an HTTP 204 No Content response is returned.

The following files were modified:
- UserController.cs: Added a new DELETE endpoint for deleting users.
- Program.cs: Updated the UpdateDatabase method to include database migration.
- UserService.cs: Modified the GetAsync method to retrieve users from the repository instead of using AutoBogus. Also added logic to handle null values when retrieving individual users.